requirements.txt
django>=5.0
celery>=5.3
redis>=5.0
python-dateutil>=2.9
psycopg2-binary>=2.9
gunicorn>=21.2